pyproject.toml
[project]
name = "elasticsearch-mcp-server"
version = "2.0.0"
description = "MCP Server for interacting with Elasticsearch"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "elasticsearch>=8.0.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=0.4.0",
    "anthropic>=0.49.0",
    "tomli>=2.2.1",
    "tomli-w>=1.2.0",
    "click>=8.1.8",
]

[project.license]
file = "LICENSE"

[project.scripts]
elasticsearch-mcp-server = "src.server:elasticsearch_mcp_server"

[tool.hatch.build.targets.wheel]
packages = [
    "src",
]

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"
